+++ packages/fuselage/src/components/SidebarV2/SidebarItem/SidebarItemCol.tsx
@@ -1,0 +1,13 @@
+import type { HTMLAttributes } from 'react';
+
+export const SidebarItemCol = ({
+  className,
+  ...props
+}: HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={['rcx-box rcx-box--full rcx-sidebar-v2-item__col', className]
+      .filter(Boolean)
+      .join(' ')}
+    {...props}
+  />
+);